Make testApi dokkaConfiguration available outside of AbstractCoreTest
Fix rebasing issue
